[FIX] html_builder: only escape text nodes on save of view
Steps to reproduce:
- Open website builder in translate mode
- Insert literally "<br>" in the text
- Save
- Bug: text "<br>" has been replaced by "&lt;br&gt;"

Commit that re-introduced the escape of text nodes:
2bee90e898cc6d6844f9005c39d862259dcd2757

task-4367641
